refactor(formatPie): add entityType and move component
